requirements.txt
mutagen>=1.47.0
musicbrainzngs>=0.7.1
pyacoustid>=1.3.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
rich>=13.7.0
